Use the newly available API plots data instead of sudo JSON
https://issues.redhat.com/browse/COST-4619
